+++ Rotation/CirceCore/Context/ICirceContext.cs
@@ -1,5 +1,6 @@
 using Olympus.Rotation.Common;
 using Olympus.Rotation.CirceCore.Helpers;
+using Olympus.Services.Party;
 
 namespace Olympus.Rotation.CirceCore.Context;
 
@@ -259,6 +260,15 @@ public interface ICirceContext : ICasterDpsRotationContext
     /// Debug state for this rotation.
     /// </summary>
     CirceDebugState Debug { get; }
+
+    #endregion
+
+    #region Party Coordination
+
+    /// <summary>
+    /// Party coordination service for raid buff synchronization.
+    /// </summary>
+    IPartyCoordinationService? PartyCoordinationService { get; }
 
     #endregion
 }