+++ Rotation/KratosCore/Context/KratosContext.cs
@@ -8,6 +8,7 @@ using Olympus.Services;
 using Olympus.Services.Action;
 using Olympus.Services.Cache;
 using Olympus.Services.Debuff;
+using Olympus.Services.Party;
 using Olympus.Services.Positional;
 using Olympus.Services.Prediction;
 using Olympus.Services.Resource;
@@ -109,6 +110,9 @@ public sealed class KratosContext : IKratosContext
     public KratosPartyHelper PartyHelper { get; }
     public KratosDebugState Debug { get; }
 
+    // Party Coordination
+    public IPartyCoordinationService? PartyCoordinationService { get; }
+
     #endregion
 
     private readonly IBattleChara? _currentTarget;
@@ -148,6 +152,7 @@ public sealed class KratosContext : IKratosContext
         bool isAtFlank,
         bool targetHasPositionalImmunity,
         ITimelineService? timelineService = null,
+        IPartyCoordinationService? partyCoordinationService = null,
         IPluginLog? log = null)
     {
         Player = player;
@@ -168,6 +173,7 @@ public sealed class KratosContext : IKratosContext
         PlayerStatsService = playerStatsService;
         TargetingService = targetingService;
         TimelineService = timelineService;
+        PartyCoordinationService = partyCoordinationService;
         ObjectTable = objectTable;
         PartyList = partyList;
         Log = log;