+++ Rotation/ThanatosCore/Context/IThanatosContext.cs
@@ -1,5 +1,6 @@
 using Olympus.Rotation.Common;
 using Olympus.Rotation.ThanatosCore.Helpers;
+using Olympus.Services.Party;
 
 namespace Olympus.Rotation.ThanatosCore.Context;
 
@@ -168,6 +169,15 @@ public interface IThanatosContext : IMeleeDpsRotationContext
     /// Debug state for this rotation.
     /// </summary>
     ThanatosDebugState Debug { get; }
+
+    #endregion
+
+    #region Party Coordination
+
+    /// <summary>
+    /// Party coordination service for raid buff synchronization.
+    /// </summary>
+    IPartyCoordinationService? PartyCoordinationService { get; }
 
     #endregion
 }