+++ Rotation/ThanatosCore/Context/ThanatosContext.cs
@@ -8,6 +8,7 @@ using Olympus.Services;
 using Olympus.Services.Action;
 using Olympus.Services.Cache;
 using Olympus.Services.Debuff;
+using Olympus.Services.Party;
 using Olympus.Services.Positional;
 using Olympus.Services.Prediction;
 using Olympus.Services.Resource;
@@ -106,6 +107,9 @@ public sealed class ThanatosContext : IThanatosContext
     public ThanatosPartyHelper PartyHelper { get; }
     public ThanatosDebugState Debug { get; }
 
+    // Party Coordination
+    public IPartyCoordinationService? PartyCoordinationService { get; }
+
     #endregion
 
     private readonly IBattleChara? _currentTarget;
@@ -146,6 +150,7 @@ public sealed class ThanatosContext : IThanatosContext
         bool isAtFlank,
         bool targetHasPositionalImmunity,
         ITimelineService? timelineService = null,
+        IPartyCoordinationService? partyCoordinationService = null,
         IPluginLog? log = null)
     {
         Player = player;
@@ -166,6 +171,7 @@ public sealed class ThanatosContext : IThanatosContext
         PlayerStatsService = playerStatsService;
         TargetingService = targetingService;
         TimelineService = timelineService;
+        PartyCoordinationService = partyCoordinationService;
         ObjectTable = objectTable;
         PartyList = partyList;
         Log = log;