+++ Rotation/ThanatosCore/Modules/BuffModule.cs
@@ -102,11 +102,40 @@ public sealed class BuffModule : IThanatosModule
             return false;
         }
 
+        // Party coordination: Synchronize with other Olympus instances
+        var partyCoord = context.PartyCoordinationService;
+        if (partyCoord != null && partyCoord.IsPartyCoordinationEnabled &&
+            context.Configuration.PartyCoordination.EnableRaidBuffCoordination)
+        {
+            // Check if our buffs are aligned with remote instances
+            // If significantly desynced (e.g., death recovery), use independently
+            if (!partyCoord.IsRaidBuffAligned(RPRActions.ArcaneCircle.ActionId))
+            {
+                context.Debug.BuffState = "Raid buffs desynced, using independently";
+                // Fall through to execute - don't try to align when heavily desynced
+            }
+            // Check if another DPS is about to use a raid buff
+            // If so, align our burst with theirs
+            else if (partyCoord.HasPendingRaidBuffIntent(
+                context.Configuration.PartyCoordination.RaidBuffAlignmentWindowSeconds))
+            {
+                context.Debug.BuffState = "Aligning with party burst";
+                // Fall through to execute and announce our intent
+            }
+
+            // Announce our intent to use Arcane Circle
+            partyCoord.AnnounceRaidBuffIntent(RPRActions.ArcaneCircle.ActionId);
+        }
+
         // Use Arcane Circle
         if (context.ActionService.ExecuteOgcd(RPRActions.ArcaneCircle, player.GameObjectId))
         {
             context.Debug.PlannedAction = RPRActions.ArcaneCircle.Name;
             context.Debug.BuffState = "Activating Arcane Circle";
+
+            // Notify coordination service that we used the raid buff
+            partyCoord?.OnRaidBuffUsed(RPRActions.ArcaneCircle.ActionId, 120_000);
+
             return true;
         }